Update calc_accuracy.py
I had to cast each 'phone' in the phone1List as a list because there is a leading \t char in the string comparision. Once I removed that in line 15 I joined the characters back into strings for comparsion. I maintained a wer of 17.0 in a string equality comparison without sorting.